chore: initial refactor pass for session inspector
